+++ packages/merkleize-metadata/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Update dependencies
+
 ## 1.1.13 - 2025-01-24
 
 ### Fixed